Handle navigation from push notifications in Test app (#2336)
