+++ DemoApp/Resources/Assets.xcassets/AppIcon.appiconset/40.png
