+++ DemoApp/Resources/Assets.xcassets/AppIcon.appiconset/60.png
