+++ DemoApp/Resources/Assets.xcassets/AppIcon.appiconset/appstore.png
