+++ DemoApp/Resources/Assets.xcassets/Contents.json
