+++ DemoApp/Resources/Assets.xcassets/Icon_arrow_right.imageset/Contents.json
