+++ DemoApp/Resources/Assets.xcassets/Stream_logo.imageset/Contents.json
