+++ DemoApp/Resources/Assets.xcassets/Stream_logo.imageset/STREAM MARK.pdf
