+++ DemoApp/Resources/Assets.xcassets/advanced_settings.imageset/Contents.json
