+++ DemoApp/Resources/Assets.xcassets/advanced_settings.imageset/advanced_icon.pdf
