+++ DemoApp/Resources/Assets.xcassets/pencil.imageset/Contents.json
