+++ DemoApp/Resources/Assets.xcassets/pencil.imageset/Icon_pen-write.pdf
