+++ DemoApp/Resources/cs.lproj/Localizable.strings
