+++ DemoApp/Resources/cs.lproj/Localizable.stringsdict
