+++ DemoApp/Shared/ChatUser+CustomFields.swift
