+++ StreamChatUITestsApp/StreamChat/DemoAppCoordinator+TestApp.swift
@@ -1,0 +1,35 @@
+//
+// Copyright © 2022 Stream.io Inc. All rights reserved.
+//
+
+import StreamChat
+import StreamChatUI
+import UIKit
+
+// MARK: - Navigation
+
+extension DemoAppCoordinator {
+    
+    func start(cid: ChannelId? = nil) {
+        if let cid = cid {
+            navigateToChannel(with: cid)
+        } else {
+            let viewController = ViewController()
+            let navigationController = UINavigationController(rootViewController: viewController)
+            set(rootViewController: navigationController, animated: false)
+        }
+    }
+    
+    private func navigateToChannel(with cid: ChannelId) {
+        if let channelList = self.window.rootViewController as? ChannelList {
+            channelList.router.showChannel(for: cid)
+        } else {
+            let viewController = ViewController()
+            let navigationController = UINavigationController(rootViewController: viewController)
+            set(rootViewController: navigationController, animated: false)
+            viewController.didTap()
+            viewController.router?.showChannel(for: cid)
+        }
+    }
+    
+}
